# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API SELETIVO",
        "version": "0.1"
    },
    "paths": {
...
        "/api/unidade": {
            "get": {
                "tags": [
                    "Unidade"
                ],
                "summary": "GET Unidade",
                "description": "GET Paginado",
                "operationId": "2250b14b758c20a6048832f40e52f176",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page Number",
                        "required": false
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {}
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/store-unidade": {
            "post": {
                "tags": [
                    "Unidade"
                ],
                "summary": "Cria uma unidade no sistema",
                "description": "Endpoint para cadastrar uma unidade",
                "operationId": "dd4a6369c27e07e1d52a49d4db7003de",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "st_data_admissao",
                                    "st_data_demissao"
                                ],
                                "properties": {
                                    "unid_nome": {
                                        "type": "string",
                                        "example": "Unidade do CPA II"
                                    },
                                    "unid_sigla": {
                                        "type": "string",
                                        "example": "GTCPAII"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Vínculo realizado com sucesso!",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Vínculo realizado com sucesso!"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Erro na validação dos dados",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Erro de validação"
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/show-unidade/{unidade_id}": {
            "get": {
                "tags": [
                    "Unidade"
                ],
                "summary": "Show Unidade",
                "description": "Show Unidade",
                "operationId": "f19c2ab254d5f8fc040ee81885d326de",
                "parameters": [
                    {
                        "name": "unidade_id",
                        "in": "path",
                        "description": "ID da Unidade",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Unidade encontrada",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "404": {
                        "description": "Nenhuma unidade encontrda"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/update-unidade/{unidade_id}": {
            "put": {
                "tags": [
                    "Unidade"
                ],
                "summary": "Atualizar uma unidade",
                "description": "Endpoint para atualizar unidade",
                "operationId": "ae0c4e35b11ebb53adfddf7e446aca4b",
                "parameters": [
                    {
                        "name": "unidade_id",
                        "in": "path",
                        "description": "ID da Unidade",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": false,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "unid_nome": {
                                        "type": "string",
                                        "example": "Unidade do CPA II"
                                    },
                                    "unid_sigla": {
                                        "type": "string",
                                        "example": "GTCPAII"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Unidade atualizado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unidade atualizado"
                                        },
                                        "data": {
                                            "properties": {
                                                "unid_nome": {
                                                    "type": "string",
                                                    "example": "Unidade do CPA II"
                                                },
                                                "unid_sigla": {
                                                    "type": "string",
                                                    "example": "GTCPAII"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Erro na requisição"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/delete-unidade/{unidade_id}": {
            "delete": {
                "tags": [
                    "Unidade"
                ],
                "summary": "Deleta Unidade",
                "description": "Remove Unidade",
                "operationId": "2aba1468e393fb612c359786f5f4a7c6",
                "parameters": [
                    {
                        "name": "unidade_id",
                        "in": "path",
                        "description": "ID da uniade",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Unidade Removida",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "404": {
                        "description": "Unidade não encontrada"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}